устранение условной компиляции для типа int, и будущее вынесение в другой проект

+++ src/hash.c
@@ -1,7 +1,7 @@
-#include "dict_couple.h"
+#include <stddef.h>
 
-int hash_func(unsigned char* text){
-    int hash_value = 0;
+size_t hash_func(unsigned char* text){
+    size_t hash_value = 0;
     while((*text) != 0){
         hash_value = (hash_value * 31 + (*text)) % 1000000007;
         text++;
